Username Avail
If signed in, moving to the username route askes the user for a username. Every key press starts the checkAvailability function if the half second timer isnt ongoing. After the half-second timer finished, the database on firestore is read for that username and the availibity is signaled eithe true or false depending on if it is found.