Photo Upload
A logged in user is prompted to upload a photo. One can be selected from their local file system, and once picked, is storage in our Firebase Storage and it's download URL linked to their user document. Also implemented rules in Storage that match the file path of the uploading picture with the uid of the user making the upload request, making sure they match ensures someone is only uploading to their own profile.