User Display Query
The page.ts runs when the page.svelte is loaded. This one takes the username typed into the URL parameters, queries the firebase user collections and returns the user info to the svelte component as a 'data' object. This data can then be used in the svelte component for frontend display