Deployed On Firebase
Deployed live on Firebase. Had to make some changes, change 1 being quite major, to get it to work. 1) Deploying to Firebase uses Firebase-Webframeworks. Firebase-Webframeworks apparently depends on Firebase version 9.9.0. I was using latest, e.g. 10.1.0, which apparently is not supported. So I downgraded the version of Firebase to 9.9.0 . 2) Added config data to firebase.json for min & max instances as deploying automatically set the instances to max by defaults but defaults also wouldnt allow them to be set that high. I set an arbitrary min/max of 0 & 5 respectively, no idea what would be a good number or what these instances even are. I also set the memory to 1GiB because someone else online had that in their config, too. Idk if necessary. 3) Updated .gitignore to ignore .firebase directory and it's firebase-debug.log 4) deleted the .VSCode file in root because it was made on accident to begin with