Login to Dashboard
After authenticating, user idToken is checked to see if has custom claim 'profile' set to true. If not, nothing happens. If true, then user has completed user profile creation (i.e. they chose a username) and should be redirected to their dashboard. The POST to the sign-in api relays this status in the response json where a if-statement in the client redirects if response.profile is true.